requirements.txt
python-decouple
django-livereload-server
django-bootstrap-v5
Pillow
django-bootstrap-icons